Added filesystem test; clean up afp errors on xattrs check

+++ entrypoint.sh
@@ -462,28 +462,39 @@ else
   fi
 fi
 
-# perform quick test to see if xattrs are supported on a path found in smb.conf
-echo "INFO: running test for xattr support on your time machine persistent storage location..."
-
-# get the first path found in smb.conf
-TEST_PATH="$(grep -m 1 "path =" < /etc/samba/smb.conf | awk '{print $3}')"
-if [ -z "${TEST_PATH}" ]
+# only run test if running samba version
+if [ -z "${NETATALK_VERSION}" ]
 then
-  echo "WARNING: unable to get test path from smb.conf; unable to test for xattr support"
-else
-  # execute test by touching a file and trying to set xattrs to the file, capture the result, and remove the test file
-  touch "${TEST_PATH}/xattr-test" || echo "WARNING: unable to write test file (is your persistent storage location read only or an invalid path?)"
-  TEST_RESULT="$(setfattr -n user.test -v "hello" "${TEST_PATH}/xattr-test" >/dev/null 2>&1; echo $?)"
-  rm -f "${TEST_PATH}/xattr-test"
+  # perform quick test to see if xattrs are supported on a path found in smb.conf
+  echo "INFO: running test for xattr support on your time machine persistent storage location..."
 
-  # check to see what the results of the set xattrs test was
-  if [ "${TEST_RESULT}" = "0" ]
+  # get the first path found in smb.conf
+  TEST_PATH="$({ grep -m 1 "path =" < /etc/samba/smb.conf | awk '{print $3}' ; } 2>/dev/null)"
+  if [ -z "${TEST_PATH}" ]
   then
-    echo "INFO: xattr test successful - your persistent data store supports xattrs"
+    echo "WARNING: unable to get test path from smb.conf; unable to test for xattr support"
   else
-    echo "WARNING: xattr test failure - unable to set xattrs on your persistent data store. Time machine backups may fail!"
+    # execute test by touching a file and trying to set xattrs to the file, capture the result, and remove the test file
+    touch "${TEST_PATH}/xattr-test" || echo "WARNING: unable to write test file (is your persistent storage location read only or an invalid path?)"
+    TEST_RESULT="$(setfattr -n user.test -v "hello" "${TEST_PATH}/xattr-test" >/dev/null 2>&1; echo $?)"
+    rm -f "${TEST_PATH}/xattr-test"
+
+    # check to see what the results of the set xattrs test was
+    if [ "${TEST_RESULT}" = "0" ]
+    then
+      echo "INFO: xattr test successful - your persistent data store supports xattrs"
+    else
+      echo "WARNING: xattr test failure - unable to set xattrs on your persistent data store. Time machine backups may fail!"
+    fi
   fi
 fi
+
+# output filesystem types detected
+for DIR in /opt/*
+do
+  DETECTED_FS="$(df -TP "${DIR}" | grep -v ^Filesystem | awk '{print $2}')"
+  echo "INFO: Detected filesystem for ${DIR} is ${DETECTED_FS}"
+done
 
 # run CMD
 echo "INFO: entrypoint complete; executing '${*}'"
